# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Courtpit API",
    "description": "Courtpit REST API.\n\n**Community.** Every request is scoped to one community. Native clients send `X-Courtpit-Community: <slug>`; browsers are resolved from the host (`<slug>.courtpit.app` or a registered custom domain). The header wins over the host. A request without a resolvable community answers `400`, an unknown one `404`.\n\n**Authentication.** A session is an opaque token, sent as `Authorization: Bearer <token>` (native) or as the `courtpit_session` cookie (web; ask for it at sign-in with `X-Courtpit-Client: web`). Either one authenticates.\n\n**Errors.** Every error is `{ \"error\": { \"code\", \"message\" } }`; clients switch on `code`.\n\n**Lists** are cursor-paginated: `?cursor=&limit=` in, `{ items, next_cursor }` out.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/matches": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Lists your matches (newest first), or a league's, or (admins) the whole community's.",
        "operationId": "list_matches",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Only matches in this status.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MatchStatus"
            }
          },
          {
            "name": "league_id",
            "in": "query",
            "description": "All matches of a league (visible to every member).",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "all",
            "in": "query",
            "description": "Admins only: every match of the community, not just your own.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor from a previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Proposes a friendly match to other members. Friendly matches never count for rankings.",
        "operationId": "create_match",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/matches/{id}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "One match with its scheduling proposals.",
        "operationId": "get_match",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/matches/{id}/cancel": {
      "post": {
        "tags": [
          "matches"
        ],
        "summary": "Calls a match off before it is played. Players may cancel friendlies; league and\ntournament matches only an admin.",
        "operationId": "cancel_match",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelMatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchView"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The body is malformed or breaks a rule.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error; see `error.code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelMatch": {
        "type": "object",
        "description": "Body of `POST /matches/{id}/cancel`.",
        "properties": {
          "note": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional reason, recorded as the resolution note."
          }
        },
        "additionalProperties": false
      },
...
      "CreateMatch": {
        "type": "object",
        "description": "Body of `POST /matches`.",
        "required": [
          "discipline",
          "opponent_ids"
        ],
        "properties": {
          "discipline": {
            "$ref": "#/components/schemas/Discipline",
            "description": "Singles, doubles or mixed."
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional place of the first proposal."
          },
          "opponent_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Side B: one player for singles, two for doubles and mixed."
          },
          "partner_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Your partner (doubles and mixed only). You play on side A."
          },
          "proposed_time": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Optional first proposal of time and place."
          }
        },
        "additionalProperties": false
      },
...
      "Deuce": {
        "type": "string",
        "description": "How a game is decided at 40–40.",
        "enum": [
          "advantage",
          "golden_point"
        ]
      },
      "Discipline": {
        "type": "string",
        "description": "What kind of tennis is played. Ranked separately (see `mixed_pooling` in scoring).",
        "enum": [
          "singles",
          "doubles",
          "mixed"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "What went wrong."
          }
        }
      },
...
      "FinalSet": {
        "type": "string",
        "description": "How the deciding set (the last possible one) is played.",
        "enum": [
          "full_set",
          "match_tiebreak_10",
          "pro_set_8"
        ]
      },
...
      "MatchFormat": {
        "type": "object",
        "description": "The rules of a match. Stored as JSON on communities, leagues and matches.",
        "required": [
          "sets_to_win",
          "games_per_set",
          "final_set"
        ],
        "properties": {
          "deuce": {
            "$ref": "#/components/schemas/Deuce",
            "description": "How games are decided at deuce. Scores are reported per set, so this does not change\nwhich scores are legal."
          },
          "final_set": {
            "$ref": "#/components/schemas/FinalSet",
            "description": "How the deciding set is played."
          },
          "games_per_set": {
            "type": "integer",
            "format": "int32",
            "description": "Games needed to win a set (usually 6).",
            "minimum": 0
          },
          "sets_to_win": {
            "type": "integer",
            "format": "int32",
            "description": "Sets needed to win: 1, 2 (best of 3) or 3 (best of 5).",
            "minimum": 0
          },
          "tiebreak_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Play a tiebreak at `n–n`. Must equal `games_per_set`; `None` means advantage sets.",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "MatchStatus": {
        "type": "string",
        "description": "Lifecycle status of a match: the flat tag of a [`MatchState`], as stored and serialized.",
        "enum": [
          "proposed",
          "scheduled",
          "reported",
          "confirmed",
          "disputed",
          "resolved",
          "walkover",
          "cancelled"
        ]
      },
...
      "Page_MatchView": {
        "type": "object",
        "description": "A page of results with an opaque cursor for the next page.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "API view of a match.",
              "required": [
                "id",
                "discipline",
                "side_a",
                "side_b",
                "side_a_names",
                "side_b_names",
                "status",
                "match_format",
                "created_at"
              ],
              "properties": {
                "confirm_deadline_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time",
                  "description": "Unanswered reports auto-confirm at this time."
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "When the row was created."
                },
                "discipline": {
                  "$ref": "#/components/schemas/Discipline",
                  "description": "Singles, doubles or mixed."
                },
                "dispute_note": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Why the reported score was disputed."
                },
                "disputed_by": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid",
                  "description": "Player who disputed the reported score."
                },
                "division_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid",
                  "description": "Division the match belongs to, if any."
                },
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "description": "Match id."
                },
                "league_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid",
                  "description": "League the match belongs to, if any."
                },
                "location": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Free-text place of play."
                },
                "match_format": {
                  "$ref": "#/components/schemas/MatchFormat",
                  "description": "Format the match is played in."
                },
                "proposals": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/Proposal"
                  },
                  "description": "Scheduling history; included on single-match responses only."
                },
                "reported_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time",
                  "description": "When the score was reported."
                },
                "reported_by": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid",
                  "description": "Player who reported the score."
                },
                "resolution_note": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Note left by the resolver."
                },
                "resolved_by": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid",
                  "description": "Player who resolved or cancelled the match."
                },
                "round": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int32",
                  "description": "Round within the league or tournament, if any."
                },
                "scheduled_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time",
                  "description": "Agreed start time, if scheduled."
                },
                "score": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/Score",
                      "description": "Reported score, if any."
                    }
                  ]
                },
                "side_a": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": "Player ids on side A."
                },
                "side_a_names": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerName"
                  },
                  "description": "Side A with display names, in the order of `side_a`."
                },
                "side_b": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": "Player ids on side B."
                },
                "side_b_names": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerName"
                  },
                  "description": "Side B with display names, in the order of `side_b`."
                },
                "status": {
                  "$ref": "#/components/schemas/MatchStatus",
                  "description": "Current lifecycle status."
                },
                "tournament_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid",
                  "description": "Tournament the match belongs to, if any."
                },
                "winner_side": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/Side",
                      "description": "Winning side, once decided."
                    }
                  ]
                }
              }
            },
            "description": "Items on this page."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass as `cursor` to fetch the next page; absent on the last page."
          }
        }
      },
...
      "PlayerName": {
        "type": "object",
        "description": "A player named by another resource (a match side, a box line, an entry, a match request),\nso clients can show names without fetching each player.",
        "required": [
          "id"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Name shown to other members; null when the viewer may not see the player (they left,\nwere banned or have not verified their email)."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Player id."
          }
        }
      },
...
      "Proposal": {
        "type": "object",
        "description": "A time-and-place proposal for a match.",
        "required": [
          "id",
          "proposed_by",
          "proposed_time",
          "status",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the proposal was made."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Proposal id."
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "description": "Proposed place, if any."
          },
          "proposed_by": {
            "type": "string",
            "format": "uuid",
            "description": "Player who made the proposal."
          },
          "proposed_time": {
            "type": "string",
            "format": "date-time",
            "description": "Proposed start time."
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus",
            "description": "Where the proposal stands."
          }
        }
      },
      "ProposalStatus": {
        "type": "string",
        "description": "Postgres `proposal_status`.",
        "enum": [
          "open",
          "accepted",
          "declined",
          "superseded"
        ]
      },
...
      "Score": {
        "type": "object",
        "description": "A reported score.",
        "required": [
          "sets"
        ],
        "properties": {
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetScore"
            },
            "description": "The sets in the order they were played."
          }
        },
        "additionalProperties": false
      },
...
      "Side": {
        "type": "string",
        "description": "One side of a match. Doubles sides have two players; any of them acts for the side.",
        "enum": [
          "a",
          "b"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token (native clients)."
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "courtpit_session",
        "description": "Session cookie (web clients); alternative to the bearer token."
      }
    }
  },
...
}